requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
pydantic[email]==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
websockets==12.0
firebase-admin==6.2.0
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
celery==5.3.4
flower==2.0.1
slowapi==0.1.9
python-multipart==0.0.6
